backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
google-genai
python-dotenv
